--- screeners.py
@@ -1,180 +1,0 @@
-from __future__ import annotations
-
-"""Screen stock data for specific criteria.
-
-This module reads OHLCV data from the ``data`` directory (generated by
-``download_ohlcv.py``) and produces two screening results:
-
-1. **setup screener** based on volume and momentum criteria.
-2. **Mark Minervini screener** based on moving average and trend rules.
-
-Results are saved in the ``result`` directory as both CSV and JSON
-files.
-"""
-
-import os
-import subprocess
-from datetime import datetime, timedelta
-from typing import Any
-
-import pandas as pd
-
-
-def ensure_pykrx() -> None:
-    """Ensure the ``pykrx`` package is installed."""
-    try:
-        import pykrx  # type: ignore  # noqa: F401
-    except ImportError:
-        print("pykrx not found. Attempting to install...")
-        subprocess.check_call(["python3", "-m", "pip", "install", "pykrx"])
-
-
-def load_market_caps(date: datetime) -> dict[str, float]:
-    """Return market caps for all tickers on ``date``."""
-    from pykrx import stock
-
-    df = stock.get_market_cap_by_ticker(date.strftime("%Y%m%d"))
-    caps = df["시가총액"].to_dict()
-    return {str(k): float(v) for k, v in caps.items()}
-
-
-def setup_screener(df: pd.DataFrame, cap: float) -> bool:
-    """Return ``True`` if ``df`` satisfies the setup screener criteria."""
-    if df.empty:
-        return False
-
-    price = df["종가"].iloc[-1]
-    if price <= 1500:
-        return False
-
-    if cap <= 60_000_000_000:
-        return False
-
-    volume = df["거래량"].iloc[-1]
-    if volume <= 100_000:
-        return False
-
-    vol60 = df["거래량"].rolling(60).mean().iloc[-1]
-    if vol60 <= 250_000:
-        return False
-
-    perf1w = df["종가"].pct_change(5).iloc[-1]
-    if not (-0.25 < perf1w < 0.25):
-        return False
-
-    perf1m = df["종가"].pct_change(21).iloc[-1]
-    if perf1m <= 0.10:
-        return False
-
-    adr = ((df["고가"] - df["저가"]) / df["종가"]).rolling(20).mean().iloc[-1] * 100
-    if adr <= 3:
-        return False
-
-    ema10 = df["종가"].ewm(span=10).mean().iloc[-1]
-    ema20 = df["종가"].ewm(span=20).mean().iloc[-1]
-    ema50 = df["종가"].ewm(span=50).mean().iloc[-1]
-
-    if not (ema50 < ema20 < ema10):
-        return False
-
-    ma20 = df["종가"].rolling(20).mean().iloc[-1]
-    std20 = df["종가"].rolling(20).std().iloc[-1]
-    bb_upper = ma20 + 2 * std20
-    if bb_upper <= price:
-        return False
-
-    return True
-
-
-def relative_strength(df: pd.DataFrame) -> float:
-    """Compute a simple 52-week relative strength measure."""
-    last_price = df["종가"].iloc[-1]
-    high52 = df["종가"].rolling(252).max().iloc[-1]
-    low52 = df["종가"].rolling(252).min().iloc[-1]
-    if high52 == low52:
-        return 0.0
-    return 100 * (last_price - low52) / (high52 - low52)
-
-
-def minervini_screener(df: pd.DataFrame) -> bool:
-    """Return ``True`` if ``df`` meets Mark Minervini criteria."""
-    if len(df) < 252:
-        return False
-
-    price = df["종가"].iloc[-1]
-
-    ma50 = df["종가"].rolling(50).mean()
-    ma150 = df["종가"].rolling(150).mean()
-    ma200 = df["종가"].rolling(200).mean()
-
-    conds = [
-        price > ma150.iloc[-1] and price > ma200.iloc[-1],
-        ma150.iloc[-1] > ma200.iloc[-1],
-        ma200.iloc[-1] > ma200.shift(21).iloc[-1],
-        ma50.iloc[-1] > ma150.iloc[-1] and ma50.iloc[-1] > ma200.iloc[-1],
-        price > ma50.iloc[-1],
-    ]
-    if not all(conds):
-        return False
-
-    low52 = df["종가"].rolling(252).min().iloc[-1]
-    high52 = df["종가"].rolling(252).max().iloc[-1]
-
-    if price <= low52 * 1.30:
-        return False
-
-    if price < high52 * 0.75:
-        return False
-
-    if relative_strength(df) < 70:
-        return False
-
-    return True
-
-
-def run() -> None:
-    """Execute both screeners and save results."""
-    ensure_pykrx()
-
-    from pykrx import stock
-
-    today = datetime.today()
-    caps = load_market_caps(today)
-
-    setup_rows: list[dict[str, Any]] = []
-    minervini_rows: list[dict[str, Any]] = []
-
-    if not os.path.isdir("data"):
-        print("data directory not found. Run download_ohlcv.py first.")
-        return
-
-    for file in os.listdir("data"):
-        if not file.endswith(".csv"):
-            continue
-        ticker = os.path.splitext(file)[0]
-        path = os.path.join("data", file)
-        df = pd.read_csv(path, index_col=0, parse_dates=True)
-        if df.empty:
-            continue
-
-        cap = caps.get(ticker, 0.0)
-
-        if setup_screener(df, cap):
-            setup_rows.append({"ticker": ticker, "relative_strength": relative_strength(df)})
-
-        if minervini_screener(df):
-            minervini_rows.append({"ticker": ticker, "relative_strength": relative_strength(df)})
-
-    os.makedirs("result", exist_ok=True)
-
-    setup_df = pd.DataFrame(setup_rows).sort_values("relative_strength", ascending=False)
-    setup_df.to_csv(os.path.join("result", "setup_screener.csv"), index=False)
-    setup_df.to_json(os.path.join("result", "setup_screener.json"), orient="records", force_ascii=False)
-
-    minervini_df = pd.DataFrame(minervini_rows).sort_values("relative_strength", ascending=False)
-    minervini_df.to_csv(os.path.join("result", "minervini_screener.csv"), index=False)
-    minervini_df.to_json(os.path.join("result", "minervini_screener.json"), orient="records", force_ascii=False)
-
-
-if __name__ == "__main__":
-    run()
